+++ changelog.md
@@ -1,5 +1,11 @@
 # PortableStructs.jl Change Log
 
+## Unreleased
+
+* Added support for construction from non-conrete types, like `UnitRange` (instead of `UnitRange{Int64}`).
+* Removed (now unnecessary) extension for `Random`.
+* Added compatibility for `OrderedCollections` v2.0.
+
 ## v1.2.0
 
 * Added support for enums in nested modules. All enums are now written with explicit module paths like `ModuleA.SubModule1.Fruits.mango`.